requirements.txt
appdirs==1.4.4
asgiref==3.2.10
asn1crypto==1.4.0
Beaker==1.11.0
Brlapi==0.8.0
CacheControl==0.12.6
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.2.1
distlib==0.3.1
distro==1.5.0
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.12.1
filelock==3.0.12
greenlet==0.4.17
html5lib==1.1
idna==2.10
louis==3.15.0
Mako==1.1.3
Markdown==3.3
MarkupSafe==1.1.1
misaka==2.1.1
msgpack==1.0.0
numpy==1.19.2
ordered-set==4.0.2
packaging==20.4
pep517==0.9.1
Pillow==8.0.1
ply==3.11
profilehooks==1.12.0
progress==1.5
psutil==5.7.2
pycparser==2.20
PyGObject==3.38.0
pynvim==0.4.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyserial==3.4
python-libtorrent==1.2.10
pytz==2020.1
requests==2.24.0
resolvelib==0.5.1
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
tk==0.1.0
toml==0.10.2
urllib3==1.25.10
virtualenv==20.0.31
webencodings==0.5.1
Werkzeug==1.0.1
